+++ translations/md-editor_en.ts
@@ -2173,18 +2173,33 @@ date: [date]
 <context>
     <name>OutlinePanel</name>
     <message>
-        <location filename="../src/widgets/outlinepanel.cpp" line="239"/>
-        <location filename="../src/widgets/outlinepanel.cpp" line="242"/>
+        <location filename="../src/widgets/outlinepanel.cpp" line="269"/>
+        <location filename="../src/widgets/outlinepanel.cpp" line="272"/>
         <source>Esquema</source>
         <translation>Outline</translation>
     </message>
     <message>
-        <location filename="../src/widgets/outlinepanel.cpp" line="244"/>
+        <location filename="../src/widgets/outlinepanel.cpp" line="274"/>
         <source>Encabezados del documento; actívalos para saltar a esa sección.</source>
         <translation>Document headings; activate one to jump to that section.</translation>
     </message>
     <message>
-        <location filename="../src/widgets/outlinepanel.cpp" line="282"/>
+        <location filename="../src/widgets/outlinepanel.cpp" line="300"/>
+        <source>Filtrar encabezados…</source>
+        <translation>Filter headings…</translation>
+    </message>
+    <message>
+        <location filename="../src/widgets/outlinepanel.cpp" line="305"/>
+        <source>Expandir todo</source>
+        <translation>Expand all</translation>
+    </message>
+    <message>
+        <location filename="../src/widgets/outlinepanel.cpp" line="308"/>
+        <source>Plegar todo</source>
+        <translation>Collapse all</translation>
+    </message>
+    <message>
+        <location filename="../src/widgets/outlinepanel.cpp" line="370"/>
         <source>Sin encabezados</source>
         <translation>No headings</translation>
     </message>